Enhancement #18: Activity classes moved to separate folder
